Teoricamente arreglado el problema, pero da un error en un caso particular, el cual hay que arreglar

+++ src/utils/Semaforo.h
@@ -18,7 +18,7 @@ public:
     int wait(); // decrementa
     int signal(); // incrementa
     void eliminar ();
-    int add(int value); //incrementa en value
+    int add(short value); //incrementa en value
     void barrier();
 
 };
